+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Features:
+* Added CPU name and frequency detection support on SPARC. (CPU, Linux)
+
 # 2.68.1
 
 Changes: